split src/images.py and src/draw.py
